App testing add, app.js modularise, entries/metrics fix
