+++ .env
@@ -1,1 +1,6 @@
 PHYSIO_APP_DB_URL='mongodb+srv://fpb:pHmFk9DVbNQA05hJ@physioapi.0l6mgks.mongodb.net/physio?retryWrites=true&w=majority'
+PORT=4001
+# ENC_KEY is a randomizer on an encryption function, makes it different to someone else's app.
+ENC_KEY=blahblahblahSomeRandomStringThatIsLongSoThatItIsSuperSecureAndHardToCrackOrGuess
+# ENC_IV is a variable that is an offset to a random generator.
+ENC_IV=SuperSecretStartingPointForSecurityGenerationStuffThatShouldBeUniqueToYourApp
